requirements.txt
fastai==2.5.3
matplotlib==3.5.1
Pillow==9.1.0
pydicom==2.3.0
streamlit==1.8.1
